+++ tests/schema/test_same_origin_reprocess/type_fruit.py
@@ -1,0 +1,6 @@
+import strawberry
+
+
+@strawberry.type
+class Fruit:
+    name: str